Twitter メッセージ投稿可能インターフェース実装

+++ BotManager.Service.Twitter/Messaging/TwitterMessage.cs
@@ -10,24 +10,24 @@ namespace BotManager.Service.Twitter.Messaging
 {
     internal class TwitterMessage : IDeletableMessage
     {
-        private readonly Tweet tweet;
+        private readonly Tweet? tweet;
         private readonly ITwitterServiceClient client;
 
-        public TwitterMessage(Tweet tweet, ITwitterServiceClient client)
+        public TwitterMessage(ITwitterServiceClient client, Tweet? tweet)
         {
             this.tweet = tweet;
             this.client = client;
         }
 
-        public string Content => tweet.Text ?? string.Empty;
+        public string Content => tweet?.Text ?? string.Empty;
 
-        public DateTime ReceivedTime => tweet.CreatedAt ?? DateTime.MinValue;
+        public DateTime ReceivedTime => tweet?.CreatedAt ?? DateTime.MinValue;
 
         public bool IsBot => false;
 
         public async Task Delete()
         {
-            if(tweet.ID is not null)
+            if(tweet is not null && tweet.ID is not null)
                 await client.DeleteTweet(tweet.ID);
         }
     }
